Extract the workstation session payload into a partial
Increment 2l of the endpoint-monolith decomposition. Moves the session bootstrap
payload builder (BuildSessionPayloadAsync) and its display-name/role/environment/
workspace mapping helpers (BuildDisplayName, BuildRole, MapEnvironment,
MapWorkspace) out of src/Meridian.Ui.Shared/Endpoints/WorkstationEndpoints.cs
into a new sibling partial WorkstationEndpoints.Session.cs.

Behavior-preserving verbatim relocation: closed call graph (the 5 methods have
zero references outside WorkstationEndpoints.cs; the only entry is the inline
GetWorkstationSession route lambda in core). The shared BuildRunDigest helper
(in the StrategyBriefing partial) and BuildRunNotes (shared) remain in core and
are reached across the partial.

Core WorkstationEndpoints.cs drops 5,451 -> 5,385 lines; ratchet baseline
retightened.
